Partially reconfigured navigation container
