+++ App.js
@@ -11,8 +11,6 @@ import { GlobalContext } from "./GlobalContext";
 import Default from "./components/Default";
 import Map from "./Map";
 import TestLogin from "./components/TestLogin";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import {
   isInEastArchitecture,
   isInInstructionalCenter,
@@ -62,7 +60,6 @@ export default function App() {
       [insights, setInsights],
     ],
   };
-  const Stack = createNativeStackNavigator();
 
   TaskManager.defineTask(LOCATION_TRACKING, async ({ data, error }) => {
     if (error) {
@@ -174,16 +171,7 @@ export default function App() {
   return (
     <GlobalContext.Provider value={globalObj}>
       <SafeAreaProvider>
-        <PaperProvider>
-          <NavigationContainer independent={true}>
-            <Stack.Navigator screenOptions={{ headerShown: false }}>
-              <Stack.Screen name="Dashboard" component={Dashboard} />
-              <Stack.Screen name="Home" component={Default} />
-              <Stack.Screen name="Login" component={TestLogin} />
-              <Stack.Screen name="Map" component={Map} />
-            </Stack.Navigator>
-          </NavigationContainer>
-        </PaperProvider>
+        <PaperProvider></PaperProvider>
       </SafeAreaProvider>
     </GlobalContext.Provider>
   );
