+++ components/InsightNavigator.js
@@ -1,0 +1,14 @@
+import { NavigationContainer } from "@react-navigation/native";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+
+export default InsightNavigator = () => {
+  const Stack = createNativeStackNavigator();
+
+  return (
+    <NavigationContainer independent={true}>
+      <Stack.Navigator screenOptions={{ headerShown: false }}>
+        <Stack.Screen name="Default" component={Default} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
